+++ PasswordGenerator.Tests/PassphraseTests.cs
@@ -1,0 +1,65 @@
+using System;
+using System.Linq;
+using NUnit.Framework;
+
+namespace PasswordGenerator.Tests
+{
+    public class PassphraseTests
+    {
+        [Test]
+        public void WordCountForEntropy_DerivesEnoughWordsToMeetTarget()
+        {
+            // 7,776-word list => ~12.925 bits/word; with the trailing number 6 words clears 80 bits,
+            // without it 7 are needed.
+            Assert.That(PassphraseGenerator.WordCountForEntropy(80, includeNumber: true), Is.EqualTo(6));
+            Assert.That(PassphraseGenerator.WordCountForEntropy(80, includeNumber: false), Is.EqualTo(7));
+        }
+
+        [Test]
+        public void WordCountForEntropy_NeverReturnsLessThanOne()
+        {
+            Assert.That(PassphraseGenerator.WordCountForEntropy(0), Is.EqualTo(1));
+            Assert.That(PassphraseGenerator.WordCountForEntropy(-50), Is.EqualTo(1));
+        }
+
+        [Test]
+        public void ForPassphraseWithEntropy_MeetsOrExceedsTarget()
+        {
+            var generator = Password.ForPassphraseWithEntropy(80);
+            Assert.That(generator.EstimateEntropyBits(), Is.GreaterThanOrEqualTo(80));
+        }
+
+        [Test]
+        public void ForPassphraseWithEntropy_ProducesDerivedWordCount()
+        {
+            var generator = Password.ForPassphraseWithEntropy(80, separator: '.', includeNumber: true);
+            var parts = generator.Next().Split('.');
+            Assert.That(parts.Length, Is.EqualTo(7)); // 6 words + trailing number
+        }
+
+        [Test]
+        public void EntropyFloor_RejectsWeakConfiguration()
+        {
+            Assert.Throws<ArgumentException>(() => Password.ForPassphrase(words: 2, minimumEntropyBits: 80));
+        }
+
+        [Test]
+        public void EntropyFloor_AllowsStrongConfiguration()
+        {
+            Assert.DoesNotThrow(() => Password.ForPassphrase(words: 8, minimumEntropyBits: 80));
+        }
+
+        [Test]
+        public void EntropyFloor_ZeroMeansNoEnforcement()
+        {
+            Assert.DoesNotThrow(() => Password.ForPassphrase(words: 1, minimumEntropyBits: 0));
+        }
+
+        [Test]
+        public void EstimateEntropyBits_IsAvailableThroughInterface()
+        {
+            IPasswordGenerator generator = Password.ForPassphrase(6);
+            Assert.That(generator.EstimateEntropyBits(), Is.GreaterThan(0));
+        }
+    }
+}